+++ src/android/GeofencePlugin.java
@@ -2,12 +2,18 @@ package com.cowbell.cordova.geofence;
 
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.util.Log;
+import android.Manifest;
+
+import com.google.gson.annotations.Expose;
 
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CordovaWebView;
+import org.apache.cordova.PermissionHelper;
+import org.apache.cordova.PluginResult;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -21,6 +27,21 @@ public class GeofencePlugin extends CordovaPlugin {
     private Context context;
     protected static Boolean isInBackground = true;
     public static CordovaWebView webView = null;
+
+    private class Action {
+        public String action;
+        public JSONArray args;
+        public CallbackContext callbackContext;
+
+        public Action(String action,JSONArray args,CallbackContext callbackContext) {
+            this.action = action;
+            this.args = args;
+            this.callbackContext = callbackContext;
+        }
+    }
+    
+    //FIXME: what about many executedActions at once
+    private Action executedAction;
 
     /**
      * @param cordova
@@ -39,9 +60,10 @@ public class GeofencePlugin extends CordovaPlugin {
 
     @Override
     public boolean execute(String action, JSONArray args,
-            CallbackContext callbackContext) throws JSONException {
+                           CallbackContext callbackContext) throws JSONException {
         Log.d(TAG, "GeofencePlugin execute action: " + action + " args: "
                 + args.toString());
+        executedAction = new Action(action, args, callbackContext);
 
         if (action.equals("addOrUpdate")) {
             List<GeoNotification> geoNotifications = new ArrayList<GeoNotification>();
@@ -66,14 +88,18 @@ public class GeofencePlugin extends CordovaPlugin {
                     .getWatched();
             callbackContext.success(Gson.get().toJson(geoNotifications));
         } else if (action.equals("initialize")) {
-            callbackContext.success();
+            initialize(callbackContext);
         } else if (action.equals("deviceReady")) {
             deviceReady();
-            callbackContext.success();
         } else {
             return false;
         }
+
         return true;
+    }
+
+    public boolean execute(Action action) throws JSONException {
+        return execute(action.action, action.args, action.callbackContext);
     }
 
     private GeoNotification parseFromJSONObject(JSONObject object) {
@@ -104,6 +130,51 @@ public class GeofencePlugin extends CordovaPlugin {
         } else {
 			intent.removeExtra("geofence.notification.data");
             webView.sendJavascript(js);
+        }
+    }
+
+    private void initialize(CallbackContext callbackContext) throws JSONException {
+        String[] permissions = {
+                Manifest.permission.ACCESS_COARSE_LOCATION,
+                Manifest.permission.ACCESS_FINE_LOCATION
+        };
+
+        if (!hasPermissions(permissions)) {
+            PermissionHelper.requestPermissions(this, 0, permissions);
+        } else {
+            callbackContext.success();
+        }
+    }
+
+    private boolean hasPermissions(String[] permissions) {
+        for (String permission : permissions) {
+            if (!PermissionHelper.hasPermission(this, permission)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    public void onRequestPermissionResult(int requestCode, String[] permissions,
+                                          int[] grantResults) throws JSONException {
+        PluginResult result;
+
+        if (executedAction != null) {
+            for (int r:grantResults) {
+                if (r == PackageManager.PERMISSION_DENIED) {
+                    Log.d(TAG, "Permission Denied!");
+                    result = new PluginResult(PluginResult.Status.ILLEGAL_ACCESS_EXCEPTION);
+                    executedAction.callbackContext.sendPluginResult(result);
+                    executedAction = null;
+                    return;
+                }
+            }
+            Log.d(TAG, "Permission Granted!");
+
+            //try to execute method once again
+            execute(executedAction);
+            executedAction = null;
         }
     }
 }